// web-ui/packages/lib-logger/src/errors/error-like.d.ts
export type ErrorLike = {
  message: string;
  name: string;
  stack?: string;
  cause?: unknown;
  source?: string;
  line?: number;
  column?: number;
};